Add: FIFO debug function and relevant typedef
Add the debug function to read some values in the registers. The newly pop-up problem is when I use the debug function in the main.c, wherever the place I put it , no matter in the while(1) loop or outside of it of in it but with delay(100), the "Debug" of the ST-LINK will block at the function "I2C_WaitOnFlagUntilTimeout"(where I put a breakpoint), showing that the "Timeout" as the maximum value(2^32), which truly confusing.