Fix: Finally we got the "FIFO FULL" signal.
The FIFO must be reset before use(refer lps22hh-5.2), and the WTM must be set before the switching on of FIFO(refer an5209-9.2.1)
Now the problem is why the "FIFO FULL" massage is 0.2Hz while the ODR is 25Hz and the WTM is 25, and how to get the data(maybe via function "FIFO_Get_Data").
By the way, I've never thought that 5s could be so long.